third_party/blink/web_tests/http/tests/security/link-crossorigin-stylesheet-use-credentials-expected.txt

CONSOLE ERROR: Access to CSS stylesheet at 'http://localhost:8080/security/resources/green-background-allow-star.php?1' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Origin' header in the response must not be the wildcard '*' when the request's credentials mode is 'include'.
CONSOLE ERROR: Access to CSS stylesheet at 'http://localhost:8080/security/resources/green-background-allow-star.php?2' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Origin' header in the response must not be the wildcard '*' when the request's credentials mode is 'include'.
Test that a linked stylesheet with a crossorigin=use-credentials attributes loads expected CORS enabled resources.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 
PASS 
PASS 
PASS 
PASS successfullyParsed is true

TEST COMPLETE
